requirements.txt
# Core data science stack
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.14.1

# Visualization
matplotlib==3.9.2
seaborn==0.13.2

# NBA data API
nba_api==1.4.1

# Logging & utilities
loguru==0.7.2
joblib==1.4.2
tqdm==4.67.1
pyyaml==6.0.2
invoke==2.2.0

# Experiment tracking
mlflow==2.16.2

# Model explainability
shap==0.46.0

# Documentation (optional)
sphinx==7.1.2

# Dev tools (optional)
flake8==7.1.1
mypy==1.11.2

# Testing (optional)
pytest==9.0.2

# CLI utilities
click==8.1.7

# Gradient boosting
xgboost==2.1.1

mlflow-skinny

requests
